infrastructure/data_pipeline/requirements.txt
faker>=24.0
psycopg2-binary>=2.9
clickhouse-connect>=0.7
kafka-python>=2.0
boto3>=1.34
